Root folder configuration (#17)
* Launch downloader automatically on container startup, avoid unnecessary restarts

This changes the supervisord configuration to start downloader
automatically when the docker container is started. This way any
completed transfers will start downloading immediately.

The downloader may exit immediately without an error if there are no
completed transfers waiting to be downloaded. There is no need for
supervisord to restart it, so let it know exit in less than 1 second is
ok.

* Fix parent id option for files list

* Add PUTIO_ROOT for parent directory

Set PUTIO_ROOT environment or configuration variable to keep automator
under a specified folder on put.io.